moved mock_call_sequence to helpertest
