+++ helpertest/mock_call_sequence.go
@@ -1,4 +1,4 @@
-package util
+package helpertest
 
 import (
 	"context"
